LayoutTests/platform/ios/imported/w3c/web-platform-tests/visual-viewport/viewport-unscaled-size-iframe-expected.txt

Viewport: Size in iframe - no page scale

Test Description: This test checks that window.visualViewport returns correct sizes without any pinch-zoom page scale applied but with scrollbars.


frames[0].window.visualViewport width and height is (200, 300).

frames[0].window.visualViewport width and height when scrollbars are present is (200, 300).


PASS window.visualViewport.width of iframe viewport should match iframe width.
PASS window.visualViewport.height of iframe viewport should match iframe height.
FAIL window.visualViewport.width of iframe viewport should not include scrollbar. assert_equals: expected 185 but got 200
FAIL window.visualViewport.height of iframe viewport should not include scrollbar. assert_equals: expected 285 but got 300
